Turn off toggle for category items + code refactoring for Header
